Moved utility function to References
